Update Temperature source files and RPi scripts Feb 6: 	- clean up reading function in temperature files 	- fix 'reading' data type from double to int 	- add read_input.py script for reading serial input from Arduinos

+++ main/TemperatureSensor.h
@@ -15,11 +15,11 @@ public:
 private:
 	// Attributes
 	int tempPin = -1; // the pin to be used for the temperature sensor
-	double singleReading;
+	int singleReading;
 	double avgReading;
 	ValidationHelper validationHelper;
 	// Functions
-	double convertReadingToUsable(double reading);
+	double convertReadingToUsable(int reading);
 };
 
 #endif
